Add support for monitoring

+++ controllers/workflow/workflow.go
@@ -9,47 +9,51 @@ import (
 	"github.com/fnikolai/frisbee/controllers/common/selector"
 	"github.com/fnikolai/frisbee/controllers/common/selector/service"
 	"github.com/pkg/errors"
-	"k8s.io/apimachinery/pkg/util/runtime"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
-func (r *Reconciler) schedule(ctx context.Context, w *v1alpha1.Workflow) (ctrl.Result, error) {
-	r.Logger.Info("Workflow Started", "name", w.GetName(), "time", time.Now())
+func (r *Reconciler) scheduleActions(topCtx context.Context, obj *v1alpha1.Workflow) {
+	ctx, cancel := context.WithCancel(topCtx)
+	defer cancel()
 
-	if action := w.Spec.Actions[len(w.Spec.Actions)-1]; action.ActionType != "Wait" {
-		return common.Failed(ctx, w, errors.New("All experiments must end with a wait function"))
-	}
+	r.Logger.Info("Workflow Started", "name", obj.GetName(), "time", time.Now())
 
-	for _, action := range w.Spec.Actions {
+	var err error
+
+	for _, action := range obj.Spec.Actions {
 		r.Logger.Info("Process Action", "type", action.ActionType, "name", action.Name, "depends", action.Depends)
 
 		switch action.ActionType {
 		case "Wait": // Wait command will block the entire controller
-			if err := r.wait(ctx, w, action.Wait); err != nil {
-				return common.Failed(ctx, w, errors.Wrapf(err, "wait %s failed", action.Name))
-			}
+			err = r.wait(ctx, obj, *action.Wait)
 
 		case "ServiceGroup":
-			go r.createServiceGroup(ctx, w, action)
+			err = r.createServiceGroup(ctx, obj, action)
 
 		case "Stop":
-			go r.stop(ctx, w, action)
+			err = r.stop(ctx, obj, action)
 
 		default:
-			return common.Failed(ctx, w, errors.Errorf("unknown action %s", action.ActionType))
+			err = errors.Errorf("unknown action %s", action.ActionType)
+		}
+
+		if err != nil {
+			common.Failed(ctx, obj, errors.Wrapf(err, "action %s failed", action.Name))
+
+			return
 		}
 	}
 
-	return common.Success(ctx, w)
+	common.Success(ctx, obj)
 }
 
-func (r *Reconciler) wait(ctx context.Context, w *v1alpha1.Workflow, spec *v1alpha1.WaitSpec) error {
+func (r *Reconciler) wait(ctx context.Context, w *v1alpha1.Workflow, spec v1alpha1.WaitSpec) error {
 	if len(spec.Complete) > 0 {
-		return common.WaitLifecycle(ctx, w.GetUID(), &v1alpha1.ServiceGroup{}, spec.Complete, v1alpha1.Complete)
+		common.WaitLifecycle(ctx, w.GetUID(), &v1alpha1.ServiceGroup{}, spec.Complete, v1alpha1.Complete)
 	}
 
 	if len(spec.Running) > 0 {
-		return common.WaitLifecycle(ctx, w.GetUID(), &v1alpha1.ServiceGroup{}, spec.Running, v1alpha1.Running)
+		common.WaitLifecycle(ctx, w.GetUID(), &v1alpha1.ServiceGroup{}, spec.Running, v1alpha1.Running)
 	}
 
 	if spec.Duration != nil {
@@ -64,57 +68,76 @@ func (r *Reconciler) wait(ctx context.Context, w *v1alpha1.Workflow, spec *v1alp
 	return nil
 }
 
-func (r *Reconciler) createServiceGroup(ctx context.Context, w *v1alpha1.Workflow, action v1alpha1.Action) {
-	if action.Depends != nil {
-		if err := r.wait(ctx, w, action.Depends); err != nil {
-			runtime.HandleError(err)
-
-			return
-		}
-	}
-
+func (r *Reconciler) createServiceGroup(ctx context.Context, obj *v1alpha1.Workflow, action v1alpha1.Action) error {
 	group := v1alpha1.ServiceGroup{}
-	group.Namespace = w.GetNamespace()
-	group.Name = action.Name
 	action.ServiceGroup.DeepCopyInto(&group.Spec)
 
-	if err := common.SetOwner(w, &group); err != nil {
-		runtime.HandleError(errors.Wrapf(err, "unable to set owner forgroup %s", group.GetName()))
+	if err := common.SetOwner(obj, &group, action.Name); err != nil {
+		return errors.Wrapf(err, "ownership failed %s", group.GetName())
+	}
 
-		return
+	if action.Depends != nil {
+		if err := r.wait(ctx, obj, *action.Depends); err != nil {
+			return errors.Wrapf(err, "dependencies failed. action %s ", action.Name)
+		}
 	}
 
 	_, err := ctrl.CreateOrUpdate(ctx, r.Client, &group, func() error { return nil })
 	if err != nil {
-		runtime.HandleError(errors.Wrapf(err, "unable to create group %s", group.GetName()))
-
-		return
+		return errors.Wrapf(err, "unable to create group %s", group.GetName())
 	}
+
+	// TODO: fix it with respect to other threads
+	// common.UpdateLifecycle(ctx, obj, &v1alpha1.ServiceGroup{}, group.GetName())
+
+	return nil
 }
 
-func (r *Reconciler) stop(ctx context.Context, w *v1alpha1.Workflow, action v1alpha1.Action) {
-	if action.Depends != nil {
-		if err := r.wait(ctx, w, action.Depends); err != nil {
-			runtime.HandleError(err)
-
-			return
-		}
-	}
-
+func (r *Reconciler) stop(ctx context.Context, obj *v1alpha1.Workflow, action v1alpha1.Action) error {
+	// Resolve affected services
 	var services []v1alpha1.Service
-
 	if action.Stop.Macro != "" {
-		services = service.Select(ctx, selector.ExpandMacro(action.Stop.Macro))
+		services = service.Select(ctx, selector.ParseMacro(action.Stop.Macro))
 	}
 
-	for i := 0; i < len(services); i++ {
-		// Change service Phase to Chaos so to ignore the failure caused by the following deletion.
-		_, _ = common.Chaos(ctx, &services[i])
+	if len(services) == 0 {
+		return errors.Errorf("no service to stop. macro %s", action.Stop.Macro)
+	}
 
-		if err := r.Client.Delete(ctx, &services[i]); err != nil {
-			runtime.HandleError(errors.Wrapf(err, "unable to delete service %s", services[i].GetName()))
-
-			return
+	if action.Depends != nil {
+		if err := r.wait(ctx, obj, *action.Depends); err != nil {
+			return errors.Wrapf(err, "dependencies failed. action %s ", action.Name)
 		}
 	}
+
+	// Without Schedule
+	if action.Stop.Schedule == nil {
+		for i := 0; i < len(services); i++ {
+			// Change service Phase to Chaos so to ignore the failure caused by the following deletion.
+			_, _ = common.Chaos(ctx, &services[i])
+
+			if err := r.Client.Delete(ctx, &services[i]); err != nil {
+				return errors.Wrapf(err, "unable to delete service %s", services[i].GetName())
+			}
+		}
+
+		return nil
+	}
+
+	// With Schedule
+	if action.Stop.Schedule != nil {
+		r.Logger.Info("Yield with Schedule", "services", services)
+
+		for service := range common.YieldByTime(ctx, action.Stop.Schedule.Cron, services...) {
+			if err := r.Client.Delete(ctx, service); err != nil {
+				return errors.Wrapf(err, "unable to delete service %s", service.GetName())
+			}
+		}
+
+		r.Logger.Info("Scheduled operations done ")
+
+		return nil
+	}
+
+	return nil
 }
